Moved FileExplorer to a separate component
TODO: fix error on create
